Правки sprint2/step1 - ограничил высоту экрана, добавил useMemo на фильтрацию массивов, добавил константу в burger-api.js, не смог разобраться с useRef - как при помщи него сделать плавный скролл до секции
