+++ src/components/BurgerConstructor/ComponentsConstructor/components-constructor.module.css
@@ -4,7 +4,6 @@
     align-items: center;
     gap: 16px;
     width: 568px;
-    height: 656px;
 }
 
 .constructorElemScroll {
