--- src/components/BurgerIngredients/BurgerTabs/burger-tabs.module.css
@@ -1,4 +1,0 @@
-.burgerMain {
-    display: flex;
-    margin-top: 20px;
-}
