--- src/components/BurgerIngredients/Ingridients/ingridients.module.css
@@ -1,61 +1,0 @@
-.ingridients {
-    z-index: 0;
-    display: flex;
-    flex-direction: column;
-    width: 600px;
-    height: 756px;
-    overflow-y: scroll;
-    border-radius: 0px;
-    margin-top: 20px;
-    overflow-x: hidden;
-}
-
-.ingridients::-webkit-scrollbar {
-    width: 8px;
-    background: #2F2F37;
-}
-
-.ingridients::-webkit-scrollbar-thumb {
-    left: 0px;
-    right: 0px;
-    top: 0px;
-    bottom: 172px;
-    background: #8585AD;
-    border: 3px solid #8585AD;
-}
-
-.mainBurgerColumn {
-    display: flex;
-    justify-content: center;
-    width: 600px;
-    margin-top: 20px;
-    flex-direction: column;
-}
-
-.mainColumn {
-    display: flex;
-    justify-content: center;
-    width: 600px;
-    margin-top: 20px;
-    flex-direction: column;
-}
-
-
-.headlineBurger {
-    width: 600px;
-    height: 30px;
-    background: #131316;
-
-}
-
-.headlineText {
-    width: 600px;
-    height: 30px;
-    /* desktop/display small */
-    font-family: 'JetBrains Mono', sans-serif;
-    font-style: normal;
-    font-weight: 700;
-    font-size: 24px;
-    /* text/primary-light (standart text) */
-    color: #F2F2F3;
-}
