+++ src/components/BurgerIngredients/burger-ingridients.module.css
@@ -3,7 +3,7 @@
     display: flex;
     flex-wrap: wrap;
     width: 600px;
-    height: 912px;
+    min-height: 600px;
     justify-content: center;
     background: #131316;
 
@@ -27,4 +27,70 @@
     color: #F2F2F3;
     align-items: center;
     text-align: center;
+}
+
+.ingridients {
+    z-index: 0;
+    display: flex;
+    flex-direction: column;
+    width: 600px;
+    height: calc(100vh - 270px);
+    overflow-y: scroll;
+    border-radius: 0px;
+    margin-top: 20px;
+    overflow-x: hidden;
+}
+.burgerMain {
+    display: flex;
+    margin-top: 20px;
+}
+
+.ingridients::-webkit-scrollbar {
+    width: 8px;
+    background: #2F2F37;
+}
+
+.ingridients::-webkit-scrollbar-thumb {
+    left: 0px;
+    right: 0px;
+    top: 0px;
+    bottom: 172px;
+    background: #8585AD;
+    border: 3px solid #8585AD;
+}
+
+.mainBurgerColumn {
+    display: flex;
+    justify-content: center;
+    width: 600px;
+    margin-top: 20px;
+    flex-direction: column;
+}
+
+.mainColumn {
+    display: flex;
+    justify-content: center;
+    width: 600px;
+    margin-top: 20px;
+    flex-direction: column;
+}
+
+
+.headlineBurger {
+    width: 600px;
+    height: 30px;
+    background: #131316;
+
+}
+
+.headlineText {
+    width: 600px;
+    height: 30px;
+    /* desktop/display small */
+    font-family: 'JetBrains Mono', sans-serif;
+    font-style: normal;
+    font-weight: 700;
+    font-size: 24px;
+    /* text/primary-light (standart text) */
+    color: #F2F2F3;
 }
